council_mcp_server/pyproject.toml
[project]
name = "council-mcp-server"
version = "0.1.0"
description = "MCP Council Server — persona prompts for geopolitical council deliberation"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11,<4"
dependencies = [
    "fastmcp (>=2.14.4,<3.0.0)",
    "python-dotenv (>=1.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
